PYTHON01/py01_requirements.txt
contourpy==1.3.2
cycler==0.12.1
flake8==7.3.0
fonttools==4.60.1
kiwisolver==1.4.9
matplotlib==3.10.7
mccabe==0.7.0
numpy==2.2.6
packaging==25.0
pillow==12.0.0
pycodestyle==2.14.0
pyflakes==3.4.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
six==1.17.0
